Move the class to log package
